refactor(playback): split Watch Mode Decisions region into GhostPlaybackLogic.WatchMode.cs
